Support JSON roles in between meals.
Helpful when testing places that default to chef-solo.

This changes two things:

**BetweenMeals::Knife**
This performs actions - and we continue the default of only acting on
ruby roles... however, we let you specify a `role_type` of `json` if
you'd like

**BetweenMeals::Changes::Role**
This is designed to tell you what changed in a repo and thus tells you
about any kind of role. This is in part because
`BetweenMeals::Changeset` doesn't have a general config object one
passes to it that would be appropriate for role-type, and passing in a
random role-specific thing seems awkward. So just returning all roles
that the user touched seems useful.

Note that there is the possibility of very odd behavior if:

* The user has both JSON and ruby roles in their repo
* The user changes a JSON role
* The user is either not specifying a `role_type` or has chosen `ruby`

In this case if the application (such as taste-tester) is using
`BetweenMeals::Knife` the upload will throw an error as the `knife
role from file` will provide the wrong filename and fail.

However, I think that's reasonable because:

1. Having both kinds of roles is not a supported configuration in
general. They will override each other in unpredictable ways on a Chef
server
2. The setup is probably broken anyway.

Also rubocop fixes