Added new helper for testing whether chef-zero is running
In taste-tester's server code, we're only checking whether the state
file's port param has a process listening on it (via port_open?).
Adding a new helper (chef_zero_running?) to not only check that a
process is listening on the port, but specifically a chef-zero proc.
Note - a corresponding change will also be needed in taste-tester to
utilize this new helper within the server.