Eliminate mkdir race condition
Some automation ends up running multiple copies of between_meals. I have
observed situations where the directory didn't exist, but by the time
mkdir was called, it did, so it crashed out. This uses similar logic to
FileUtils.mkdir_p to try to create the directory in any case, accept
failures as part of normal practice, and raise errors if the directory
*didn't* get created.