Better handling for git config
Currently if `user.email` or `user.name` are not set, `git config`
returns an error. This means that if, if you're taste-tester plugin
you check to see if it's config, taste-tester just blows up deep
inside the bowels of between-meals, which is suboptimal.

Since `git config` returns 1 in the case of no config set, let's
accept that as a valid response.

Signed-off-by: Phil Dibowitz <phil@ipom.com>