Allow users to get access to the Rugged object
For the repo checks hook it's useful to get access to the rugged object
to do sanity checks without having to shell out. This provides an
accessor.

It's done specifically in the git module because it's the only module
in which we use such a library, the other ones are simple shellouts.

Signed-off-by: Phil Dibowitz <phil@ipom.com>