Fix change detection for databags
Summary:
For changes detection we don't care if file was added or modified, we only need to understand
if file was modified or deleted. So using same logic as in
BetweenMeals::Changes::Role

This would fix an issue with Grocery Delivery uploads when adding new
databag.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: